+++ src/main/java/com/mobble/mobbleserver/domain/notification/device/Platform.java
@@ -1,0 +1,6 @@
+package com.mobble.mobbleserver.domain.notification.device;
+
+public enum Platform {
+    ANDROID,
+    IOS
+}
