--- src/main/java/com/mobble/mobbleserver/refactor/notification/device/entity/Platform.java
@@ -1,6 +1,0 @@
-package com.mobble.mobbleserver.refactor.notification.device.entity;
-
-public enum Platform {
-    ANDROID,
-    IOS
-}
